requirements.txt
Flask==2.3.1
solace-pubsubplus==1.5.0